+++ app/Models/AdminProfile.php
@@ -24,6 +24,8 @@ class AdminProfile extends Model
 
     // Fillables
     protected $fillable = [
+        'profile_id',
+        'user_id',
         'last_name',
         'first_name',
         'middle_name',